Modify model to match ccd data. Moved related files to package.
