+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/model/ccd/AgreedDraftOrder.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.draftorders.upload.DraftOrderCollection;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.draftorders.upload.PensionSharingAnnexCollection;
 
 import java.util.List;
