+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/model/ccd/draftorders/upload/SuggestedDraftOrderAdditionalDocumentsCollection.java
@@ -1,0 +1,18 @@
+package uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.draftorders.upload;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseDocument;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.HasCaseDocument;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class SuggestedDraftOrderAdditionalDocumentsCollection implements HasCaseDocument {
+    private CaseDocument value;
+}
